+++ frontend/resourceadm/components/FieldsetWrapper/index.ts
@@ -1,0 +1,1 @@
+export { FieldsetWrapper } from './FieldsetWrapper';
