+++ frontend/resourceadm/components/ResourceReferenceFields/index.ts
@@ -1,0 +1,1 @@
+export { ResourceReferenceFields } from './ResourceReferenceFields';
